+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/paperswithcode/adapters/output/persistence/neo4j/Neo4jBenchmarkRepository.kt
@@ -8,11 +8,24 @@ import org.springframework.data.neo4j.repository.Neo4jRepository
 interface Neo4jBenchmarkRepository : Neo4jRepository<Neo4jResource, Long> {
 
     @Query("""
-MATCH (r:ResearchField {resource_id: {0}})<-[:RELATED {predicate_id: 'P30'}]-(p:Paper)-[:RELATED {predicate_id: 'P31'}]->(cont:Contribution)
+MATCH (:ResearchField {resource_id: {0}})-[:RELATED* {predicate_id: 'P36'}]->(f:ResearchField)
+WITH COLLECT(f) AS fields
+MATCH (f:ResearchField {resource_id: {0}})
+WITH fields + f AS fields
+UNWIND fields AS field
+MATCH (field)<-[:RELATED {predicate_id: 'P30'}]-(p:Paper)-[:RELATED {predicate_id: 'P31'}]->(cont:Contribution)
 MATCH (cont)-[:RELATED {predicate_id: 'HAS_BENCHMARK'}]->(:Benchmark)-[:RELATED {predicate_id: 'HAS_DATASET'}]->(ds:Dataset)
 MATCH (cont)-[:RELATED {predicate_id: 'P32'}]->(pr:Problem)
 OPTIONAL MATCH (cont)-[:RELATED {predicate_id: 'HAS_SOURCE_CODE'}]->(l:Literal)
-RETURN DISTINCT pr AS problem, COUNT(DISTINCT p) AS totalPapers, COUNT(DISTINCT l) AS totalCodes, COUNT(DISTINCT ds) AS totalDatasets
-    """)
+RETURN DISTINCT pr AS problem, field, COUNT(DISTINCT p) AS totalPapers, COUNT(DISTINCT l) AS totalCodes, COUNT(DISTINCT ds) AS totalDatasets    """)
     fun summarizeBenchmarkByResearchField(id: ResourceId): Iterable<Neo4jBenchmarkSummary>
+
+    @Query("""
+MATCH (f:ResearchField)<-[:RELATED {predicate_id: 'P30'}]-(p:Paper)-[:RELATED {predicate_id: 'P31'}]->(cont:Contribution)
+MATCH (cont)-[:RELATED {predicate_id: 'HAS_BENCHMARK'}]->(:Benchmark)-[:RELATED {predicate_id: 'HAS_DATASET'}]->(ds:Dataset)
+MATCH (cont)-[:RELATED {predicate_id: 'P32'}]->(pr:Problem)
+OPTIONAL MATCH (cont)-[:RELATED {predicate_id: 'HAS_SOURCE_CODE'}]->(l:Literal)
+RETURN DISTINCT pr AS problem, f AS field, COUNT(DISTINCT p) AS totalPapers, COUNT(DISTINCT l) AS totalCodes, COUNT(DISTINCT ds) AS totalDatasets  
+    """)
+    fun summarizeBenchmarkGetAll(): Iterable<Neo4jBenchmarkSummary>
 }