+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/paperswithcode/adapters/output/persistence/neo4j/Neo4jEntities.kt
@@ -5,11 +5,13 @@ import eu.tib.orkg.prototype.paperswithcode.application.domain.Dataset
 import eu.tib.orkg.prototype.paperswithcode.application.domain.DatasetSummary
 import eu.tib.orkg.prototype.researchproblem.application.domain.ResearchProblem
 import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jResource
+import eu.tib.orkg.prototype.statements.domain.model.ResearchField
 import org.springframework.data.neo4j.annotation.QueryResult
 
 @QueryResult
 data class Neo4jBenchmarkSummary(
     val problem: Neo4jResource,
+    val field: Neo4jResource,
     val totalPapers: Long,
     val totalDatasets: Long,
     val totalCodes: Long
@@ -17,6 +19,7 @@ data class Neo4jBenchmarkSummary(
     fun toBenchmarkSummary() =
         BenchmarkSummary(
             ResearchProblem(problem.resourceId!!, problem.label!!),
+            ResearchField(field.thingId!!, field.label!!),
             totalPapers.toInt(),
             totalDatasets.toInt(),
             totalCodes.toInt()