+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/paperswithcode/application/port/input/UseCases.kt
@@ -10,6 +10,7 @@ import java.util.Optional
 
 interface RetrieveBenchmarkUseCase {
     fun summariesForResearchField(id: ResourceId): Optional<List<BenchmarkSummary>>
+    fun summary(): Optional<List<BenchmarkSummary>>
 }
 
 interface RetrieveDatasetUseCase {