+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/paperswithcode/application/port/output/Queries.kt
@@ -8,6 +8,7 @@ import eu.tib.orkg.prototype.statements.domain.model.ResourceId
 
 interface SummarizeBenchmarkQuery {
     fun byResearchField(id: ResourceId): List<BenchmarkSummary>
+    fun getAll(): List<BenchmarkSummary>
 }
 
 interface FindDatasetsQuery {