+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/paperswithcode/application/service/PwcServices.kt
@@ -30,6 +30,11 @@ class BenchmarkService(
             summarizeBenchmark.byResearchField(researchField.get().id!!)
         )
     }
+
+    override fun summary(): Optional<List<BenchmarkSummary>> =
+        Optional.of(
+            summarizeBenchmark.getAll()
+        )
 }
 
 @Primary