+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Fetch benchmarks of all research fields for PwC use-case (Described in: [!240](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/240))
 
 ## [0.13.3] - 2022-08-05
 ### Changed